Newswires UI: add support for local timezone
