+++ newswires/client/package.json
@@ -24,6 +24,7 @@
 		"@emotion/react": "11.13.3",
 		"lodash": "4.17.21",
 		"moment": "2.30.1",
+		"moment-timezone": "^0.5.47",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"sanitize-html": "^2.13.0",
@@ -58,6 +59,7 @@
 		"jest-environment-jsdom": "^29.7.0",
 		"prettier": "^3.3.3",
 		"storybook": "8.6.0",
+		"timezone-mock": "^1.3.6",
 		"ts-jest": "29.2.4",
 		"typescript": "^5.5.3",
 		"typescript-eslint": "^8.0.1",
