+++ newswires/client/src/context/SearchReducer.test.ts
@@ -1,8 +1,11 @@
 import dateMath from '@elastic/datemath';
 import moment from 'moment/moment';
+import { register } from 'timezone-mock';
 import { sampleWireData } from '../tests/fixtures/wireData.ts';
 import type { Action, State } from './SearchContext.tsx';
 import { SearchReducer } from './SearchReducer';
+
+register('Etc/GMT');
 
 jest.mock('@elastic/datemath', () => ({
 	__esModule: true,
@@ -21,12 +24,18 @@ describe('SearchReducer', () => {
 	const successState: State = {
 		...initialState,
 		status: 'success',
-		queryData: { results: [sampleWireData], totalCount: 1 },
+		queryData: {
+			results: [{ ...sampleWireData, ingestedAt: '2025-01-01T00:00:00+00:00' }],
+			totalCount: 1,
+		},
 	};
 
 	const offlineState: State = {
 		status: 'offline',
-		queryData: { results: [sampleWireData], totalCount: 1 },
+		queryData: {
+			results: [{ ...sampleWireData, ingestedAt: '2025-01-01T00:00:00+00:00' }],
+			totalCount: 1,
+		},
 		successfulQueryHistory: [],
 		error: 'Network error',
 		autoUpdate: true,
@@ -34,7 +43,10 @@ describe('SearchReducer', () => {
 
 	const errorState: State = {
 		status: 'error',
-		queryData: { results: [sampleWireData], totalCount: 1 },
+		queryData: {
+			results: [{ ...sampleWireData, ingestedAt: '2025-01-01T00:00:00+00:00' }],
+			totalCount: 1,
+		},
 		successfulQueryHistory: [],
 		error: 'Fetch error',
 		autoUpdate: true,
@@ -50,7 +62,13 @@ describe('SearchReducer', () => {
 		const newState = SearchReducer(initialState, action);
 
 		expect(newState.status).toBe('success');
-		expect(newState.queryData).toEqual(action.data);
+		expect(newState.queryData?.results).toHaveLength(1);
+		expect(newState.queryData?.totalCount).toBe(1);
+		expect(newState.queryData?.results).toContainEqual({
+			...sampleWireData,
+			id: 1,
+			ingestedAt: '2025-01-01T00:00:00+00:00',
+		});
 		expect(newState.successfulQueryHistory).toEqual([
 			{ query: action.query, resultsCount: 1 },
 		]);
@@ -88,7 +106,15 @@ describe('SearchReducer', () => {
 		it(`should handle UPDATE_RESULTS action in ${state.status} state`, () => {
 			const action: Action = {
 				type: 'UPDATE_RESULTS',
-				data: { results: [{ ...sampleWireData, id: 2 }], totalCount: 1 },
+				data: {
+					results: [
+						{
+							...sampleWireData,
+							id: 2,
+						},
+					],
+					totalCount: 1,
+				},
 				query: { q: 'test' },
 			};
 
@@ -101,21 +127,23 @@ describe('SearchReducer', () => {
 				...sampleWireData,
 				id: 2,
 				isFromRefresh: true,
+				ingestedAt: '2025-01-01T00:00:00+00:00',
 			});
 			expect(newState.queryData?.results).toContainEqual({
 				...sampleWireData,
 				id: 1,
+				ingestedAt: '2025-01-01T00:00:00+00:00',
 			});
 		});
 	});
 
-	it(`should filter storing when handling UPDATE_RESULTS action in success state`, () => {
+	it(`should filter stories when handling UPDATE_RESULTS action in success state`, () => {
 		const state: State = {
 			...successState,
 			queryData: {
 				results: [
-					{ ...sampleWireData, id: 1, ingestedAt: '2025-01-01T02:00:00Z' },
-					{ ...sampleWireData, id: 2, ingestedAt: '2025-01-01T02:05:00Z' },
+					{ ...sampleWireData, id: 1, ingestedAt: '2025-01-01T02:00:00+00:00' },
+					{ ...sampleWireData, id: 2, ingestedAt: '2025-01-01T02:05:00+00:00' },
 				],
 				totalCount: 2,
 			},
@@ -140,13 +168,13 @@ describe('SearchReducer', () => {
 		expect(state.queryData.results).toContainEqual({
 			...sampleWireData,
 			id: 2,
-			ingestedAt: '2025-01-01T02:05:00Z',
+			ingestedAt: '2025-01-01T02:05:00+00:00',
 		});
 
 		expect(state.queryData.results).toContainEqual({
 			...sampleWireData,
 			id: 1,
-			ingestedAt: '2025-01-01T02:00:00Z',
+			ingestedAt: '2025-01-01T02:00:00+00:00',
 		});
 
 		const newState = SearchReducer(state, action);
@@ -158,34 +186,39 @@ describe('SearchReducer', () => {
 		expect(newState.queryData?.results).toContainEqual({
 			...sampleWireData,
 			id: 4,
-			ingestedAt: '2025-01-01T02:07:00Z',
+			ingestedAt: '2025-01-01T02:07:00+00:00',
 			isFromRefresh: true,
 		});
 
 		expect(newState.queryData?.results).toContainEqual({
 			...sampleWireData,
 			id: 3,
-			ingestedAt: '2025-01-01T02:06:00Z',
+			ingestedAt: '2025-01-01T02:06:00+00:00',
 			isFromRefresh: true,
 		});
 
 		expect(newState.queryData?.results).toContainEqual({
 			...sampleWireData,
 			id: 2,
-			ingestedAt: '2025-01-01T02:05:00Z',
+			ingestedAt: '2025-01-01T02:05:00+00:00',
 		});
 
 		expect(newState.queryData?.results).not.toContainEqual({
 			...sampleWireData,
 			id: 1,
-			ingestedAt: '2025-01-01T02:00:00Z',
+			ingestedAt: '2025-01-01T02:00:00+00:00',
 		});
 	});
 
 	it(`should handle APPEND_RESULTS action in success state`, () => {
 		const state: State = {
 			...successState,
-			queryData: { results: [{ ...sampleWireData, id: 2 }], totalCount: 2 },
+			queryData: {
+				results: [
+					{ ...sampleWireData, id: 2, ingestedAt: '2025-01-01T00:00:00+00:00' },
+				],
+				totalCount: 2,
+			},
 		};
 
 		const action: Action = {
@@ -201,10 +234,12 @@ describe('SearchReducer', () => {
 		expect(newState.queryData?.results).toContainEqual({
 			...sampleWireData,
 			id: 1,
+			ingestedAt: '2025-01-01T00:00:00+00:00',
 		});
 		expect(newState.queryData?.results).toContainEqual({
 			...sampleWireData,
 			id: 2,
+			ingestedAt: '2025-01-01T00:00:00+00:00',
 		});
 	});
