+++ newswires/client/src/dateHelpers.test.ts
@@ -4,7 +4,7 @@ import {
 	deriveDateMathRangeLabel,
 	isValidDateValue,
 	relativeDateRangeToAbsoluteDateRange,
-} from './dateMathHelpers.ts';
+} from './dateHelpers.ts';
 
 jest.mock('@elastic/datemath', () => ({
 	__esModule: true,
